stockAnalysis/requirements.txt
datetime
sqlalchemy
psycopg2-binary
requests
django-prometheus
django-cors-headers
djangorestframework-simplejwt
django_rest_framework